Pulled all HIDG interface abstraction code into a separate file
